crawler/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
APScheduler==3.10.4
python-dotenv==1.0.1
pyyaml==6.0.2
httpx==0.27.2
